Ajuste calculo percentual pergunta

+++ src/SME.SR.Infra/Dtos/Relatorios/Sondagem/RelatorioSondagemComponentesMatematicaNumerosAutoralConsolidadoRespostaDto.cs
@@ -3,7 +3,7 @@
     public class RelatorioSondagemComponentesMatematicaNumerosAutoralConsolidadoRespostaDto
     {
         public string Resposta { get; set; }
-        public int AlunosQuantidade { get; set; }
+        public double AlunosQuantidade { get; set; }
         public double AlunosPercentual { get; set; }
     }
 }
